requirements.txt
Django=5.0.4
pyserial=3.5
setuptools=66.1.1
asgiref=3.8.1
sqlparse=0.5.0

